requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
google-cloud-storage==2.15.0
google-auth==2.28.2
hashids==1.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0